The Product Description Component
In app.component.html, replace <h1>{{title}}</h1> with <app-product-page></app-product-page>

create a new component by typing ng generate component product-description

Look in product-page.component.html for the div with a class of description and move that div tag and all of its children to the product-description.component.html file, replacing any HTML that’s already in the ProductDescriptionComponent file.

In product-page.component.html in the empty space where the old div was, add an opening and closing tag for the app-product-description selector. Once you complete this task, the page should render the same in the browser, but now the HTML is being loaded from the ProductDescriptionComponent.

In product-description.component.css, select the p tag and set the font-size to 16px, the font-family to Helvetica with a fallback of Arial and sans-serif, and a font-weight of normal.